separate staking and rewards dirs
