+++ hooks/queries/rewards/useOutstandingRewards.ts
