+++ web3/index.ts
@@ -1,27 +1,27 @@
 export { createVotingContractInstance } from "./contracts/createVotingContractInstance";
 export { createVotingTokenContractInstance } from "./contracts/createVotingTokenContractInstance";
+export { removeDelegate } from "./mutations/delegation/removeDelegate";
+export { removeDelegator } from "./mutations/delegation/removeDelegator";
+export { setDelegate } from "./mutations/delegation/setDelegate";
+export { setDelegator } from "./mutations/delegation/setDelegator";
 export { approve } from "./mutations/staking/approve";
 export { executeUnstake } from "./mutations/staking/executeUnstake";
 export { requestUnstake } from "./mutations/staking/requestUnstake";
 export { stake } from "./mutations/staking/stake";
 export { withdrawAndRestake } from "./mutations/staking/withdrawAndRestake";
 export { withdrawRewards } from "./mutations/staking/withdrawRewards";
-export { removeDelegate } from "./mutations/delegation/removeDelegate";
-export { removeDelegator } from "./mutations/delegation/removeDelegator";
-export { setDelegate } from "./mutations/delegation/setDelegate";
-export { setDelegator } from "./mutations/delegation/setDelegator";
 export { commitVotes } from "./mutations/votes/commitVotes";
 export { revealVotes } from "./mutations/votes/revealVotes";
-export { getOutstandingRewards } from "./queries/balances/getOutstandingRewards";
-export { getStakerDetails } from "./queries/balances/getStakerDetails";
-export { getTokenAllowance } from "./queries/balances/getTokenAllowance";
-export { getUnstakeCoolDown } from "./queries/balances/getUnstakeCoolDown";
-export { getUnstakedBalance } from "./queries/balances/getUnstakedBalance";
 export { getDelegateSetEvents } from "./queries/delegation/getDelegateSetEvents";
 export { getDelegateToStaker } from "./queries/delegation/getDelegateToStaker";
 export { getDelegatorSetEvents } from "./queries/delegation/getDelegatorSetEvents";
 export { getIgnoredRequestToBeDelegateAddresses } from "./queries/delegation/getIgnoredRequestToBeDelegateAddresses";
 export { getVoterFromDelegate } from "./queries/delegation/getVoterFromDelegate";
+export { getOutstandingRewards } from "./queries/rewards/getOutstandingRewards";
+export { getStakerDetails } from "./queries/staking/getStakerDetails";
+export { getTokenAllowance } from "./queries/staking/getTokenAllowance";
+export { getUnstakeCoolDown } from "./queries/staking/getUnstakeCoolDown";
+export { getUnstakedBalance } from "./queries/staking/getUnstakedBalance";
 export { getActiveVotes } from "./queries/votes/getActiveVotes";
 export { getCommittedVotes } from "./queries/votes/getCommittedVotes";
 export { getEncryptedVotes } from "./queries/votes/getEncryptedVotes";
