+++ web3/queries/rewards/getOutstandingRewards.ts
