+++ web3/queries/staking/getStakerDetails.ts
